+++ ElvUI_MerathilisUI/Core/Changelog/6.44.lua
@@ -1,0 +1,12 @@
+local MER = unpack(ElvUI_MerathilisUI)
+
+MER.Changelog[644] = {
+	RELEASE_DATE = "10.11.2024",
+	FIXES = {
+		"[Announcement]: Fixed an undefined local function. Thx @DakJaniels",
+		"[BigWigs]: Use new API to register a Profile",
+		"[Event Tracker]: Disable the Radient Echo Event by default",
+		"[OmniCD]: Fix latest changes to it",
+		"[Skins]: Various tweaks",
+	},
+}