+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Changes ###
 
-  * [AutoButtons]: Fixed an issues with the AutoButtons options
-  * [EventTracker]: Fixed an time table option
+  * [Announcement]: Fixed an undefined local function. Thx @DakJaniels
+  * [BigWigs]: Use new API to register a Profile
+  * [Event Tracker]: Disable the Radient Echo Event by default
+  * [OmniCD]: Fix latest changes to it
+  * [Skins]: Various tweaks